git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/928-CFA-to-C-exporter-prototype@40330 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/smg/SMGWriteReadTest.java
@@ -977,7 +977,7 @@ public class SMGWriteReadTest extends SMGTest0 {
     SMGObject testObject = createRegion(BigInteger.valueOf(256));
     smg = smg.copyAndAddObject(testObject);
     // Write the entire SMG to the zero value in 1 Byte blocks
-    for (int i = 0; i < 32; i++) {
+    for (long i = 0; i < 32; i++) {
       smg =
           smg.writeValue(
               testObject, BigInteger.valueOf(i * 8), BigInteger.valueOf(8), SMGValue.zeroValue());
@@ -1005,7 +1005,7 @@ public class SMGWriteReadTest extends SMGTest0 {
     SMGObject testObject = createRegion(BigInteger.valueOf(256));
     smg = smg.copyAndAddObject(testObject);
     // Write the SMG to the zero value in 1 Byte blocks except every 4th Byte
-    for (int i = 0; i < 32; i++) {
+    for (long i = 0; i < 32; i++) {
       if (i % 4 != 3) {
         smg =
             smg.writeValue(
